Finish moving all old code into respective locations in new file tree
